More accurately limit number of posts returned
Reddit's API seems to be returning the limit number you request plus whatever
pinned posts there are for that subreddit. This limits in the appending loop so
that it only lists the things that are most currently active in the subreddit
(may or may not omit pinned depending on their activity).